Update patch set 1
Patch Set 1:

https://review.lineageos.org/#/c/159780/
The hint was removed in lineage because it wasn't even used. But the powerhal not handling a hint explicitly shouldn't cause build errors. There's several that aren't handled in that function. You sure you have this commit downstream and aren't getting the error that made me push this in?

Patch-set: 1